Refactor: Create separate packages for KiwixMobileActivity and KiwixErrorActivity
