+++ app/src/androidTestKiwix/java/org/kiwix/kiwixmobile/tests/activities/KiwixMobileActivityTest.java
@@ -9,7 +9,7 @@ import android.test.suitebuilder.annotation.LargeTest;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.kiwix.kiwixmobile.KiwixMobileActivity;
+import org.kiwix.kiwixmobile.main.KiwixMobileActivity;
 
 @LargeTest
 @RunWith(AndroidJUnit4.class)
