+++ app/src/androidTestKiwix/java/org/kiwix/kiwixmobile/tests/activities/SearchActivityTest.java
@@ -9,7 +9,6 @@ import android.test.suitebuilder.annotation.LargeTest;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.kiwix.kiwixmobile.KiwixMobileActivity;
 import org.kiwix.kiwixmobile.search.SearchActivity;
 
 @LargeTest
