+++ app/src/main/AndroidManifest.xml
@@ -35,7 +35,7 @@
     </activity>
 
     <activity
-      android:name=".KiwixMobileActivity"
+      android:name=".main.KiwixMobileActivity"
       android:configChanges="orientation|keyboardHidden|screenSize|locale"
       android:label="@string/app_name"
       android:windowSoftInputMode="adjustPan">
@@ -178,7 +178,7 @@
     <service android:name=".downloader.DownloadService" />
 
     <activity
-      android:name=".KiwixErrorActivity"
+      android:name=".error.KiwixErrorActivity"
       android:process=":error_activity" />
 
     <provider
