+++ app/src/main/java/org/kiwix/kiwixmobile/data/contentprovider/ZimContentProvider.java
@@ -38,7 +38,7 @@ import org.kiwix.kiwixlib.JNIKiwixString;
 import org.kiwix.kiwixlib.Pair;
 import org.kiwix.kiwixmobile.BuildConfig;
 import org.kiwix.kiwixmobile.KiwixApplication;
-import org.kiwix.kiwixmobile.KiwixMobileActivity;
+import org.kiwix.kiwixmobile.main.KiwixMobileActivity;
 import org.kiwix.kiwixmobile.utils.files.FileUtils;
 
 import java.io.File;
