+++ app/src/main/java/org/kiwix/kiwixmobile/downloader/DownloadFragment.java
@@ -41,7 +41,7 @@ import android.widget.RelativeLayout;
 import android.widget.TextView;
 
 import org.kiwix.kiwixmobile.KiwixApplication;
-import org.kiwix.kiwixmobile.KiwixMobileActivity;
+import org.kiwix.kiwixmobile.main.KiwixMobileActivity;
 import org.kiwix.kiwixmobile.R;
 import org.kiwix.kiwixmobile.base.BaseFragment;
 import org.kiwix.kiwixmobile.library.entity.LibraryNetworkEntity;
