+++ app/src/main/java/org/kiwix/kiwixmobile/error/KiwixErrorActivity.java
@@ -1,4 +1,4 @@
-package org.kiwix.kiwixmobile;
+package org.kiwix.kiwixmobile.error;
 
 import android.content.Context;
 import android.content.Intent;
@@ -10,6 +10,7 @@ import android.support.v4.content.FileProvider;
 import android.widget.Button;
 import android.widget.CheckBox;
 
+import org.kiwix.kiwixmobile.R;
 import org.kiwix.kiwixmobile.base.BaseActivity;
 import org.kiwix.kiwixmobile.data.contentprovider.ZimContentProvider;
 import org.kiwix.kiwixmobile.data.database.BookDao;
