+++ app/src/main/java/org/kiwix/kiwixmobile/settings/KiwixSettingsActivity.java
@@ -39,7 +39,7 @@ import android.widget.Toast;
 
 import org.kiwix.kiwixmobile.BuildConfig;
 import org.kiwix.kiwixmobile.KiwixApplication;
-import org.kiwix.kiwixmobile.KiwixMobileActivity;
+import org.kiwix.kiwixmobile.main.KiwixMobileActivity;
 import org.kiwix.kiwixmobile.R;
 import org.kiwix.kiwixmobile.base.BaseActivity;
 import org.kiwix.kiwixmobile.data.database.RecentSearchDao;
