+++ app/src/main/java/org/kiwix/kiwixmobile/views/KiwixActionBarDrawerToggle.java
@@ -22,7 +22,7 @@ import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
-import org.kiwix.kiwixmobile.KiwixMobileActivity;
+import org.kiwix.kiwixmobile.main.KiwixMobileActivity;
 import org.kiwix.kiwixmobile.R;
 
 public class KiwixActionBarDrawerToggle extends ActionBarDrawerToggle {
