+++ app/src/main/java/org/kiwix/kiwixmobile/views/web/ToolbarStaticKiwixWebView.java
@@ -21,7 +21,7 @@ import android.content.Context;
 import android.util.AttributeSet;
 import android.view.ViewGroup;
 
-import org.kiwix.kiwixmobile.KiwixMobileActivity;
+import org.kiwix.kiwixmobile.main.KiwixMobileActivity;
 import org.kiwix.kiwixmobile.utils.DimenUtils;
 
 /**
