+++ app/src/main/res/layout/activity_kiwix_error.xml
@@ -5,7 +5,7 @@
   android:layout_width="match_parent"
   android:layout_height="match_parent"
   android:background="@android:color/holo_blue_dark"
-  tools:context="org.kiwix.kiwixmobile.KiwixErrorActivity">
+  tools:context="org.kiwix.kiwixmobile.error.KiwixErrorActivity">
 
   <TextView
     android:id="@+id/messageText"
